refactored tulevaTarjonta
